requirements/dev.txt
# requirements/dev.txt
-r common.txt
django-debug-toolbar==1.6
ipython==5.1.0
ipython-genutils==0.1.0
